Added accelerometer based control.
Added in accelerometer control and moved the menu based system to the up button. This seems to work with the current deadzone, but may need adjustment based on the user or use case. That said, it works! IP address is still hardcoded, and is currently set to use a mobile hotspot to save me time.